Add build statuses
Builds now have a "config_extracted" field which is changed to true when
the config is extracted from the git repo. Not sure how to handle this
when there is trouble extracting the config. Statueses are stored on
each step and aggregated via Build.status/1. Most of the changes involve
updating the status in the database in the right places. This may be
rewritten later using GenEvent (something we need to do for a plugin API).